refactor: Move validator structs to lib.rs
