+++ programs/sunrise_beam/src/instructions/register_state.rs
@@ -1,28 +1,5 @@
-use crate::seeds::GSOL_MINT_AUTHORITY;
-use crate::state::{ControllerState, RegisterStateInput};
-use crate::token::create_mint;
+use crate::{state::RegisterStateInput, token::create_mint, RegisterState};
 use anchor_lang::prelude::*;
-use anchor_spl::token::Token;
-
-#[derive(Accounts)]
-#[instruction(input: RegisterStateInput)]
-pub struct RegisterState<'info> {
-    #[account(
-        init,
-        space = ControllerState::size(input.initial_capacity),
-        payer = payer
-    )]
-    pub state: Account<'info, ControllerState>,
-    #[account(mut)]
-    pub payer: Signer<'info>,
-    #[account(mut)]
-    pub gsol_mint: Signer<'info>,
-    #[account(seeds = [state.key().as_ref(), GSOL_MINT_AUTHORITY], bump)]
-    pub gsol_mint_authority: UncheckedAccount<'info>,
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token>,
-    pub rent: Sysvar<'info, Rent>,
-}
 
 pub fn handler(ctx: Context<RegisterState>, input: RegisterStateInput) -> Result<()> {
     let state_account = &mut ctx.accounts.state;
