+++ programs/sunrise_beam/src/instructions/remove_beam.rs
@@ -1,13 +1,5 @@
-use crate::state::*;
-use crate::BeamProgramError;
+use crate::{BeamProgramError, RemoveBeam};
 use anchor_lang::prelude::*;
-
-#[derive(Accounts)]
-pub struct RemoveBeam<'info> {
-    #[account(mut, has_one = update_authority)]
-    pub state: Account<'info, ControllerState>,
-    pub update_authority: Signer<'info>,
-}
 
 pub fn handler(ctx: Context<RemoveBeam>, beam: Pubkey) -> Result<()> {
     if ctx.accounts.state.remove_beam(&beam).is_none() {
